Models/VGTD/requirements.txt
multidict==6.0.2
murmurhash==1.0.7
mypy-extensions==0.4.3
nltk==3.7
numba==0.55.1
numpy==1.19.2
oauthlib==3.2.0
omegaconf==2.1.2
opt-einsum==3.3.0
optuna==2.10.0
packaging==21.3
parso==0.8.3
pathspec==0.7.0
pathtools==0.1.2
pathy==0.6.1
pbr==5.9.0
pexpect==4.8.0
pickleshare==0.7.5
pip==21.2.4
preshed==3.0.6
prettytable==3.3.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyDeprecate==0.3.1
Pygments==2.11.2
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytorch-lightning==1.4.9
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools==59.5.0
shortuuid==1.0.9
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sparse==0.13.0
SQLAlchemy==1.4.36
srsly==2.4.3
stack-data==0.2.0
stevedore==3.5.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==8.0.16
tokenizers==0.12.1
toml==0.10.2
torchmetrics==0.6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.19.2
typed-ast==1.4.3
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
wandb==0.12.16
wasabi==0.9.1
wcwidth==0.2.5
Werkzeug==2.1.2
wheel==0.37.1
yarl==1.7.2
zipp==3.8.0
